Split render::{Graph, Thread} mod from graph.rs
